+++ codes/lc/lc0683.cpp
@@ -18,8 +18,8 @@ public:
 		for (int d = 0; d < n; d++) days[bulbs[d] - 1] = d + 1;
 
 		int res = INT_MAX;
-		int left = 0, right = K + 1;
-		for (int b = 1; right < bulbs.size(); b++) {
+		size_t left = 0, right = K + 1;
+		for (size_t b = 1; right < bulbs.size(); b++) {
 			// bulb b is good, continue scanning
 			if (days[b] > days[left] && days[b] > days[right]) continue;
 
@@ -37,14 +37,14 @@ public:
 	int kEmptySlots(vector<int>& bulbs, int K) {
 		size_t n = bulbs.size();
 		vector<int> days(n); // ith bulbs is turned on on (days[i])th day.
-		for (int i = 0; i < n; i++)
+		for (size_t i = 0; i < n; i++)
 			days[bulbs[i] - 1] = i + 1;
 
 		int res = INT_MAX;
-		deque<int> win;
+		deque<size_t> win;
 		win.push_front(0);
 
-		for (int b = 1; b < n; b++) {
+		for (size_t b = 1; b < n; b++) {
 			if (win.front() + K + 1 < b) win.pop_front();
 			int f = win.front();
 			while (win.size() && days[b] < days[win.back()]) win.pop_back();
@@ -58,7 +58,7 @@ public:
 
 	int kEmptySlots1(vector<int>& bulbs, int K) {
 		set<int> days;
-		for (int i = 0; i < bulbs.size(); i++) {
+		for (size_t i = 0; i < bulbs.size(); i++) {
 			auto it = days.insert(bulbs[i]).first;
 			if (it != days.begin()) {
 				--it;
